requirements.txt
ultralytics==8.0.20
insightface==0.7.3
opencv-python==4.8.0.76
numpy==1.24.3
scikit-learn==1.3.0
Pillow==10.0.0
sqlalchemy==2.0.19
loguru==0.7.0
requests==2.31.0
PyYAML==6.0.1
Flask==2.3.3
# For DeepSort/ByteTrack (if used)
filterpy==1.4.5
lap==0.4.0
# For RTSP support
ffmpeg-python==0.2.0 